+++ src/coord_transformations/combined_mapping.hpp
@@ -20,11 +20,12 @@
  * The functions in this mapping are defined on the coordinate system associated
  * with the domain @f$ \Omega_{mid} @f$.
  */
-template <class Mapping1, class Mapping2, class CoordJacobianType = typename Mapping2::CoordResult>
+template <
+        concepts::MappingWithJacobian Mapping1,
+        concepts::Mapping Mapping2,
+        class CoordJacobianType = typename Mapping2::CoordResult>
 class CombinedMapping
 {
-    static_assert(is_mapping_v<Mapping1>);
-    static_assert(is_mapping_v<Mapping2>);
     static_assert(std::is_same_v<typename Mapping2::CoordResult, typename Mapping1::CoordArg>);
     static_assert(
             (std::is_same_v<CoordJacobianType, typename Mapping2::CoordArg>)
@@ -93,7 +94,7 @@ public:
      *
      * @return The equivalent coordinate.
      */
-    CoordResult operator()(CoordArg coord)
+    CoordResult operator()(CoordArg coord) const
     {
         return m_mapping_1(m_mapping_2(coord));
     }
@@ -120,7 +121,6 @@ public:
         if constexpr (std::is_same_v<CoordJacobian, typename Mapping2::CoordResult>) {
             static_assert(is_analytical_mapping_v<Mapping2>);
             using InverseMapping2 = inverse_mapping_t<Mapping2>;
-            static_assert(has_jacobian_v<Mapping1>);
             static_assert(has_jacobian_v<InverseMapping2>);
             static_assert(std::is_same_v<CoordJacobian, typename Mapping1::CoordJacobian>);
             static_assert(std::is_same_v<CoordJacobian, typename InverseMapping2::CoordJacobian>);