+++ src/coord_transformations/linear_coord_transform.hpp
@@ -5,39 +5,82 @@
 
 #include "tensor.hpp"
 
-template <class StartDim, class EndDim>
+/**
+ * @brief A class describing a linear coordinate transformation.
+ *
+ * A class describing a linear coordinate transformation of the form:
+ * @f$ x_1 = \alpha x_2 + \beta @f$
+ * where @f$ x_1 @f$ and @f$ x_2 @f$ are coordinates in the input and
+ * output dimensions, and @f$ \alpha @f$ and @f$ \beta @f$ are coefficients.
+ */
+template <class InputDim, class OutputDim, class CoordJacob = Coord<InputDim>>
 class LinearCoordTransform
 {
 public:
-    using CoordArg = Coord<StartDim>;
-    using CoordResult = Coord<EndDim>;
-    using CoordJacobian = CoordArg;
+    /// The type of the argument of the function described by this mapping
+    using CoordArg = Coord<InputDim>;
+    /// The type of the result of the function described by this mapping
+    using CoordResult = Coord<OutputDim>;
+    /// The type of the coordinate that can be used to evaluate the Jacobian of this mapping
+    using CoordJacobian = CoordJacob;
 
 private:
-    Coord<StartDim> m_start_reference;
-    Coord<EndDim> m_end_reference;
+    Coord<InputDim> m_reference_point_on_input_dim;
+    Coord<OutputDim> m_reference_point_on_output_dim;
     double m_scaling_factor;
 
 public:
+    /**
+     * @brief A constructor for the linear coordinate transformation.
+     *
+     * The constructor infers the linear coordinate transformation from a
+     * reference point provided in both coordinate systems and a scaling factor.
+     *
+     * The transformation is then defined as:
+     * @f$ x_{out} = x_{out}^* + s * (x_{in} - x_{in}^*) @f$
+     * where @f$ x_{in} @f$ and @f$ x_{out} @f$ denote the point expressed in the
+     * input and output coordinate system, @f$ \cdot^* @f$ denotes the reference
+     * point and @f$ s @f$ denotes the scaling factor.
+     *
+     * @param reference_point_on_input_dim
+     *              The reference point expressed in the input coordinate system.
+     * @param reference_point_on_output_dim
+     *              The reference point expressed in the output coordinate system.
+     * @param scaling_factor
+     *              The scaling factor describing how distances in the output
+     *              coordinate system scale compared to distances in the input
+     *              coordinate system.
+     */
     explicit KOKKOS_FUNCTION LinearCoordTransform(
-            Coord<StartDim> start_reference,
-            Coord<EndDim> end_reference,
+            Coord<InputDim> reference_point_on_input_dim,
+            Coord<OutputDim> reference_point_on_output_dim,
             double scaling_factor)
-        : m_start_reference(start_reference)
-        , m_end_reference(end_reference)
+        : m_reference_point_on_input_dim(reference_point_on_input_dim)
+        , m_reference_point_on_output_dim(reference_point_on_output_dim)
         , m_scaling_factor(scaling_factor)
     {
     }
 
-    KOKKOS_DEFAULTED_FUNCTION LinearCoordTransform(LinearCoordTransform const&) = default;
+    /**
+     * Copy constructor for LinearCoordTransform.
+     * @param[in] other The operator to be copied.
+     */
+    KOKKOS_DEFAULTED_FUNCTION LinearCoordTransform(LinearCoordTransform const& other) = default;
 
     KOKKOS_DEFAULTED_FUNCTION ~LinearCoordTransform() = default;
 
+    /**
+     * @brief Convert the coordinate on the input dimension to a coordinate on the output dimension.
+     *
+     * @param[in] coord The coordinate to be converted expressed on the input coordinate system.
+     *
+     * @return The coordinate expressed on the output coordinate system.
+     */
     KOKKOS_FUNCTION CoordResult operator()(CoordArg const& coord) const
     {
-        double arg_dist = coord - m_start_reference;
+        double arg_dist = coord - m_reference_point_on_input_dim;
         double res_dist = arg_dist * m_scaling_factor;
-        return m_end_reference + res_dist;
+        return m_reference_point_on_output_dim + res_dist;
     }
 
     /**
@@ -58,17 +101,17 @@ public:
      *
      * For some computations, we need the complete Jacobian matrix or just the
      * coefficients.
-     * The coefficients can be given independently with the function jacobian_component.
+     * The coefficient can be given as a scalar with the function jacobian_component.
      *
      * @param[in] coord
      * 				The coordinate where we evaluate the Jacobian matrix.
      * @return The Jacobian matrix.
      */
-    KOKKOS_FUNCTION DTensor<VectorIndexSet<EndDim>, VectorIndexSet<StartDim>> jacobian_matrix(
+    KOKKOS_FUNCTION DTensor<VectorIndexSet<OutputDim>, VectorIndexSet<InputDim>> jacobian_matrix(
             CoordArg const& coord) const
     {
-        DTensor<VectorIndexSet<EndDim>, VectorIndexSet<StartDim>> jacobian_matrix;
-        ddcHelper::get<EndDim, StartDim>(jacobian_matrix) = m_scaling_factor;
+        DTensor<VectorIndexSet<OutputDim>, VectorIndexSet<InputDim>> jacobian_matrix;
+        ddcHelper::get<OutputDim, InputDim>(jacobian_matrix) = m_scaling_factor;
         return jacobian_matrix;
     }
 
@@ -83,8 +126,8 @@ public:
     template <class IndexTag1, class IndexTag2>
     KOKKOS_FUNCTION double jacobian_component(CoordArg const& coord) const
     {
-        static_assert(std::is_same_v<IndexTag1, EndDim>);
-        static_assert(std::is_same_v<IndexTag2, StartDim>);
+        static_assert(std::is_same_v<IndexTag1, OutputDim>);
+        static_assert(std::is_same_v<IndexTag2, InputDim>);
         return m_scaling_factor;
     }
 
@@ -94,25 +137,23 @@ public:
      *
      * For some computations, we need the complete inverse Jacobian matrix or just the
      * coefficients.
-     * The coefficients can be given independently with the function inv_jacobian_component.
+     * The coefficient can be given as a scalar with the function inv_jacobian_component.
      *
      * @param[in] coord
      * 				The coordinate where we evaluate the Jacobian matrix.
      * @return The inverse Jacobian matrix.
      */
-    KOKKOS_FUNCTION DTensor<VectorIndexSet<StartDim>, VectorIndexSet<EndDim>> inv_jacobian_matrix(
-            CoordArg const& coord) const
+    KOKKOS_FUNCTION DTensor<VectorIndexSet<InputDim>, VectorIndexSet<OutputDim>>
+    inv_jacobian_matrix(CoordArg const& coord) const
     {
-        DTensor<VectorIndexSet<StartDim>, VectorIndexSet<EndDim>> matrix;
-        ddcHelper::get<StartDim, EndDim>(matrix) = 1.0 / m_scaling_factor;
+        DTensor<VectorIndexSet<InputDim>, VectorIndexSet<OutputDim>> matrix;
+        ddcHelper::get<InputDim, OutputDim>(matrix) = 1.0 / m_scaling_factor;
         return matrix;
     }
 
 
     /**
      * @brief Compute the (i,j) coefficient of the inverse Jacobian matrix.
-     *
-     * Be careful because not all mappings are invertible, especially at the centre point.
      *
      * @param[in] coord
      *              The coordinate where we evaluate the inverse Jacobian matrix.
@@ -122,8 +163,8 @@ public:
     template <class IndexTag1, class IndexTag2>
     KOKKOS_FUNCTION double inv_jacobian_component(CoordArg const& coord) const
     {
-        static_assert(std::is_same_v<IndexTag1, StartDim>);
-        static_assert(std::is_same_v<IndexTag2, EndDim>);
+        static_assert(std::is_same_v<IndexTag1, InputDim>);
+        static_assert(std::is_same_v<IndexTag2, OutputDim>);
         return 1.0 / m_scaling_factor;
     }
 
@@ -132,10 +173,11 @@ public:
      *
      * @return The inverse mapping.
      */
-    KOKKOS_INLINE_FUNCTION LinearCoordTransform<EndDim, StartDim> get_inverse_mapping() const
+    KOKKOS_INLINE_FUNCTION LinearCoordTransform<OutputDim, InputDim> get_inverse_mapping() const
     {
-        return LinearCoordTransform<
-                EndDim,
-                StartDim>(m_end_reference, m_start_reference, 1.0 / m_scaling_factor);
+        return LinearCoordTransform<OutputDim, InputDim>(
+                m_reference_point_on_output_dim,
+                m_reference_point_on_input_dim,
+                1.0 / m_scaling_factor);
     }
 };