+++ src/multipatch/interface_derivatives/single_interface_derivatives_calculator.hpp
@@ -7,6 +7,7 @@
 #include "ddc_aliases.hpp"
 #include "edge.hpp"
 #include "edge_transformation.hpp"
+#include "geometry_descriptors.hpp"
 #include "types.hpp"
 
 
@@ -248,6 +249,68 @@ public:
                 Bound1,
                 Bound2)
     {
+        static_assert(ddc::is_discrete_domain_v<IdxRangeA>);
+        static_assert(ddc::is_discrete_domain_v<IdxRangeB>);
+    }
+
+    /**
+     * @brief Instantiate SingleInterfaceDerivativesCalculator. 
+     * See @ref SingleInterfaceDerivativesCalculatorInstantiator.
+     * This constructor calculates an approximation of the formula. 
+     * @param idx_range_1d_1 1D index range perpendicular to the Interface, 
+     * on the patch 1. 
+     * @param idx_range_1d_2 1D index range perpendicular to the Interface, 
+     * on the patch 2. 
+     * @param number_chosen_cells The number of cells we choose for the 
+     * computation of the interface derivatives. Same number for Patch1 and Patch2. 
+     * @warning If the mesh has additional interpolation points, please be sure to
+     * not take these points. 
+     */
+    SingleInterfaceDerivativesCalculator(
+            IdxRange1DPerp_1 const& idx_range_1d_1,
+            IdxRange1DPerp_2 const& idx_range_1d_2,
+            std::size_t const number_chosen_cells)
+        : SingleInterfaceDerivativesCalculator(
+                (m_extremity_1 == Extremity::FRONT)
+                        ? idx_range_1d_1.take_first(IdxStep<EdgePerpGrid1>(
+                                Kokkos::min(number_chosen_cells + 1, idx_range_1d_1.size())))
+                        : idx_range_1d_1.take_last(IdxStep<EdgePerpGrid1>(
+                                Kokkos::min(number_chosen_cells + 1, idx_range_1d_1.size()))),
+                (m_extremity_2 == Extremity::FRONT)
+                        ? idx_range_1d_2.take_first(IdxStep<EdgePerpGrid2>(
+                                Kokkos::min(number_chosen_cells + 1, idx_range_1d_2.size())))
+                        : idx_range_1d_2.take_last(IdxStep<EdgePerpGrid2>(
+                                Kokkos::min(number_chosen_cells + 1, idx_range_1d_2.size()))),
+                ddc::BoundCond::HERMITE,
+                ddc::BoundCond::HERMITE)
+    {
+    }
+
+    /**
+     * @brief Instantiate SingleInterfaceDerivativesCalculator. 
+     * See @ref SingleInterfaceDerivativesCalculatorInstantiator.
+     * This constructor calculates an approximation of the formula. 
+     * @param idx_range_a Index range on one patch. 
+     * @param idx_range_b Index range on the other patch. 
+     * @param number_chosen_cells The number of cells we choose for the 
+     * computation of the interface derivatives. Same number for Patch1 and Patch2. 
+     * @warning If the mesh has additional interpolation points, please be sure to
+     * not take these points. 
+     */
+    template <class IdxRangeA, class IdxRangeB>
+    SingleInterfaceDerivativesCalculator(
+            IdxRangeA const& idx_range_a,
+            IdxRangeB const& idx_range_b,
+            std::size_t const number_chosen_cells)
+        : SingleInterfaceDerivativesCalculator(
+                IdxRange1DPerp_1(
+                        ddc::cartesian_prod_t<IdxRangeA, IdxRangeB>(idx_range_a, idx_range_b)),
+                IdxRange1DPerp_2(
+                        ddc::cartesian_prod_t<IdxRangeA, IdxRangeB>(idx_range_a, idx_range_b)),
+                number_chosen_cells)
+    {
+        static_assert(ddc::is_discrete_domain_v<IdxRangeA>);
+        static_assert(ddc::is_discrete_domain_v<IdxRangeB>);
     }
 
     /**